# backend/openapi/v1.yaml
openapi: 3.1.0
info:
  title: Image Generation Workflow API
  version: 0.1.0
  description: |
    MVP API contract for image generation job lifecycle and ComfyUI callback integration.
servers:
  - url: http://localhost:8000
paths:
  /healthz:
    get:
      tags: [system]
      summary: Health check
      responses:
        '200':
          description: Healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
  /v1/jobs:
    post:
      tags: [jobs]
      summary: Create a generation job
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobCreateRequest'
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobCreateResponse'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /v1/jobs/{job_id}:
    get:
      tags: [jobs]
      summary: Query job status
      parameters:
        - in: path
          name: job_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatusResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /v1/jobs/{job_id}/cancel:
    post:
      tags: [jobs]
      summary: Cancel a non-terminal job
      parameters:
        - in: path
          name: job_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobCancelResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /v1/uploads/images:
    post:
      tags: [uploads]
      summary: Upload source image for generation
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadImageResponse'
        '413':
          $ref: '#/components/responses/PayloadTooLargeError'
        '415':
          $ref: '#/components/responses/UnsupportedMediaTypeError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /v1/callbacks/comfyui:
    post:
      tags: [callbacks]
      summary: Receive ComfyUI callback event
      parameters:
        - in: header
          name: X-ComfyUI-Signature
          required: false
          schema:
            type: string
          description: |
            HMAC-SHA256 signature. Signed message format:
            `timestamp + '.' + nonce + '.' + raw_body`.
        - in: header
          name: X-ComfyUI-Timestamp
          required: false
          schema:
            type: string
          description: Unix timestamp in seconds.
        - in: header
          name: X-ComfyUI-Nonce
          required: false
          schema:
            type: string
          description: Random nonce for replay protection.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComfyUICallbackRequest'
      responses:
        '200':
          description: Callback accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComfyUICallbackAck'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'

components:
  schemas:
    JobStatus:
      type: string
      enum: [queued, running, postprocessing, succeeded, failed, timeout, cancelled]
    ErrorCode:
      type: string
      enum:
        - not_found
        - invalid_state_transition
        - invalid_callback
        - validation_error
        - conflict
        - unauthorized
        - unsupported_media_type
        - payload_too_large
        - internal_error
    ErrorObject:
      type: object
      required: [code, message]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
        request_id:
          type: string
          nullable: true
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorObject'
    JobCreateRequest:
      type: object
      required: [workflow_type, workflow_version]
      properties:
        workflow_type:
          type: string
          minLength: 1
          maxLength: 64
          example: model_photo_generation
        workflow_version:
          type: string
          minLength: 1
          maxLength: 32
          example: v1
        workflow_params:
          type: object
          additionalProperties: true
        priority:
          type: integer
          minimum: 1
          maximum: 10
          default: 5
        callback_url:
          type: string
          format: uri
          nullable: true
        client_reference:
          type: string
          maxLength: 128
          nullable: true
    JobCreateResponse:
      type: object
      required: [job_id, status, progress, created_at]
      properties:
        job_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/JobStatus'
        progress:
          type: number
          minimum: 0
          maximum: 100
        created_at:
          type: string
          format: date-time
        idempotency_key:
          type: string
          nullable: true
    JobStatusResponse:
      type: object
      required:
        - job_id
        - status
        - progress
        - workflow_type
        - workflow_version
        - created_at
        - updated_at
        - output_urls
      properties:
        job_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/JobStatus'
        progress:
          type: number
          minimum: 0
          maximum: 100
        workflow_type:
          type: string
        workflow_version:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        output_urls:
          type: array
          items:
            type: string
        error_message:
          type: string
          nullable: true
    JobCancelResponse:
      type: object
      required: [job_id, status, cancelled_at]
      properties:
        job_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/JobStatus'
        cancelled_at:
          type: string
          format: date-time
    ComfyUICallbackEvent:
      type: string
      enum: [accepted, progress, completed, postprocess_completed, failed]
    ComfyUICallbackRequest:
      type: object
      required: [job_id, event]
      properties:
        job_id:
          type: string
          format: uuid
        event:
          $ref: '#/components/schemas/ComfyUICallbackEvent'
        provider_request_id:
          type: string
          nullable: true
          maxLength: 128
        progress:
          type: number
          minimum: 0
          maximum: 1
          nullable: true
        output_urls:
          type: array
          items:
            type: string
        error_message:
          type: string
          nullable: true
          maxLength: 2000
        raw_payload:
          type: object
          additionalProperties: true
    ComfyUICallbackAck:
      type: object
      required: [job_id, accepted, status]
      properties:
        job_id:
          type: string
          format: uuid
        accepted:
          type: boolean
          default: true
        status:
          $ref: '#/components/schemas/JobStatus'
    UploadImageResponse:
      type: object
      required: [upload_id, file_name, content_type, size_bytes, image_url, created_at]
      properties:
        upload_id:
          type: string
          format: uuid
        file_name:
          type: string
        content_type:
          type: string
        size_bytes:
          type: integer
          minimum: 1
        image_url:
          type: string
        created_at:
          type: string
          format: date-time

  responses:
    PayloadTooLargeError:
      description: Uploaded file exceeds maximum size
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnsupportedMediaTypeError:
      description: Uploaded file media type is not allowed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: Unauthorized callback request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ConflictError:
      description: Business conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'